requirements.txt
asgiref==3.2.7
celery==4.3.0
cffi==1.14.0
cryptography==2.9
Django==3.0.5
django-rest-knox==4.1.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
gunicorn==19.9.0
psycopg2==2.8.3
pycparser==2.20
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
redis==3.3.6